Minor changes and hide dotted line when price is hidden see #178
