+++ src/blocks/menu-food-section/edit.js
@@ -3,12 +3,13 @@
  */
 
 const {Fragment} = wp.element;
-const {withSelect} = wp.data;
 
 /**
  * Internal dependencies
  */
 import FoodMenuSectionPreview from "./preview";
+
+import { withInnerBlocks } from "../../utils";
 
 const FoodMenuSectionEdit = function( props ) {
 	return (
@@ -17,17 +18,5 @@ const FoodMenuSectionEdit = function( props ) {
 		</Fragment>
 	);
 };
-
-const withInnerBlocks = withSelect( ( select, props ) => {
-	const { clientId } = props;
-	const { getBlock } = select( 'core/block-editor' );
-	const parentBlock = getBlock( clientId );
-	const innerBlocks = parentBlock.innerBlocks;
-
-	return {
-		innerBlocks,
-		...props
-	}
-} );
 
 export default withInnerBlocks( FoodMenuSectionEdit );
