+++ src/blocks/menu-food/scss/style.scss
@@ -144,6 +144,10 @@
 		height: 0;
 		line-height: 0;
 		border-bottom: 0.2em dotted currentColor;
+
+		.price--is-hidden & {
+			content: none;
+		}
 	}
 }
