+++ src/utils.js
@@ -1,5 +1,7 @@
 import {getProps, getState, getStylesFromProps} from "./components/with-parallax/util";
 
+const { withSelect } = wp.data;
+
 export const debounce = (func, wait) => {
 	let timeout = null;
 
@@ -22,7 +24,7 @@ export const range = function( min, max ) {
 		array.push( i + min );
 	}
 	return array;
-}
+};
 
 export const withFirstBlockConditions = function( Component ) {
 
@@ -64,7 +66,7 @@ export const hasTouchScreen = function() {
 	}
 
 	return hasTouchScreen;
-}
+};
 
 export const findParents = ( target, query ) => {
 	let parents = [];
@@ -108,7 +110,7 @@ export const shuffleArray = function( array ) {
 export const defaultSnapValues = {
 	x: [0, 0.5, 1],
 	y: [0, 0.5, 1]
-}
+};
 
 export const maybeSnapFocalPoint = function( focalPoint, snapValues = defaultSnapValues ) {
 	let x = parseFloat( focalPoint.x );
@@ -128,10 +130,10 @@ export const maybeSnapFocalPoint = function( focalPoint, snapValues = defaultSna
 	} );
 
 	return { x, y }
-}
+};
 
 export const getSnapClassname = function( focalPoint ) {
-	const classNames = []
+	const classNames = [];
 
 	if ( defaultSnapValues.x.includes( parseFloat( focalPoint.x ) ) ) {
 		classNames.push( 'is-snapped-x' );
@@ -142,4 +144,16 @@ export const getSnapClassname = function( focalPoint ) {
 	}
 
 	return classNames.join( ' ' );
-}
+};
+
+export const withInnerBlocks = withSelect( ( select, props ) => {
+	const { clientId } = props;
+	const { getBlock } = select( 'core/block-editor' );
+	const parentBlock = getBlock( clientId );
+	const innerBlocks = parentBlock.innerBlocks;
+
+	return {
+		innerBlocks,
+		...props
+	}
+} );
